AxisSpan: move VSpan and HSpan into distinct files
