+++ src/ScottPlot/Plottable/AxisSpan.cs
@@ -1,33 +1,10 @@
 ﻿using ScottPlot.Drawing;
-using ScottPlot.Ticks;
 using System;
 using System.Diagnostics;
 using System.Drawing;
 
 namespace ScottPlot.Plottable
 {
-    /// <summary>
-    /// Shaded horizontal region between two X values
-    /// </summary>
-    public class HSpan : AxisSpan
-    {
-        public double X1 { get => Position1; set => Position1 = value; }
-        public double X2 { get => Position2; set => Position2 = value; }
-        public HSpan() : base(true) { }
-        public override string ToString() => $"Horizontal span between Y1={X1} and Y2={X2}";
-    }
-
-    /// <summary>
-    /// Shade the region between two Y values
-    /// </summary>
-    public class VSpan : AxisSpan
-    {
-        public double Y1 { get => Position1; set => Position1 = value; }
-        public double Y2 { get => Position2; set => Position2 = value; }
-        public VSpan() : base(false) { }
-        public override string ToString() => $"Vertical span between X1={Y1} and X2={Y2}";
-    }
-
     public abstract class AxisSpan : IPlottable, IDraggable, IHasColor, IHasArea
     {
         // location and orientation
